Conditionally load config / verify SSL
For a run from Tower (i.e. paste this script into Tower), you'll need to skip
SSL verification. For a run from a shell, you can (and should) use SSL
verification - assuming your Tower is a Satellite client.

This patch checks if it runs from Tower and skips SSL verification if so. It
also reads the config from a hard-coded /etc/ansible/hammer.ini if it finds it
runs from Tower.

Behaviour from a shell invocation remains the same as before.